server/requirements.txt
django==2.2.5
djangorestframework==3.10.3
python-dateutil==2.8.0
